requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0

# Data Science Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Geospatial & Satellite Data
rasterio>=1.3.0
geopandas>=0.13.0
sentinelsat>=1.2.0
shapely>=2.0.0
pyproj>=3.5.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
folium>=0.14.0

# Dashboard
streamlit>=1.25.0
streamlit-folium>=0.13.0

# API & Web
requests>=2.31.0

# Image Processing
Pillow>=10.0.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
